backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
python-multipart==0.0.20
pyotp==2.9.0
qrcode==8.0
pyyaml==6.0.2
